+++ packages/sources/stader-address-list/src/endpoint/address.ts
@@ -439,7 +439,7 @@ export class AddressTransport extends SubscriptionTransport<EndpointTypes> {
         handler: async (index) => {
           const pageNumber = index + 1
           const addresses: string[] =
-            await params.permissionlessNodeRegistryManager.getNodeELVaultAddressForOptOutOperators(
+            await params.permissionlessNodeRegistryManager.getAllNodeELVaultAddress(
               pageNumber,
               params.batchSize,
               {