+++ packages/sources/stader-balance/src/endpoint/utils.ts
@@ -15,11 +15,13 @@ const SECONDS_PER_SLOT = 12
 const SLOTS_PER_EPOCH = 32
 
 export const WITHDRAWAL_DONE_STATUS = 'withdrawal_done'
+export const PENDING_INITIALIZED = 'pending_initialized'
 export const DEPOSIT_EVENT_TOPIC =
   '0x649bbc62d0e31342afea4e5cd82d4049e7e1ee912fc0889aa790803be39038c5'
 export const DEPOSIT_EVENT_LOOKBACK_WINDOW = 10_000 // blocks
 export const ONE_ETH_WEI = BigNumber(ethers.utils.parseEther('1').toString())
 export const THIRTY_ONE_ETH_WEI = BigNumber(ethers.utils.parseEther('31').toString())
+export const THIRTY_TWO_ETH_WEI = BigNumber(ethers.utils.parseEther('32').toString())
 
 export type NetworkChainMap = {
   [network: string]: {