+++ packages/composites/por-indexer/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @chainlink/por-indexer-adapter
 
+## 1.2.56
+
+### Patch Changes
+
+- Updated dependencies [4f4ca79e1]
+- Updated dependencies [d25ef6507]
+  - @chainlink/proof-of-reserves-adapter@1.18.0
+
 ## 1.2.55
 
 ### Patch Changes